refactor(ui): colocate search-tools into route-level _components
Moves the SearchTools component tree out of the shared src/components dump into the search-tools route segment under _components, now that the app router migration owns the route. The page imports from a trimmed ./_components barrel that exposes only what it renders

SearchToolSelector is not used by the search-tools page; its only consumers are OldTeams and team/TeamInfo, so it moves to its own shared feature folder src/components/search_tools, mirroring how the sibling VectorStoreSelector lives in vector_store_management rather than being buried in a consumer's folder

Relative ../ imports in the moved files are rewritten to @/components/* absolute paths, and the four pre-existing eslint-suppressions entries are re-keyed to the new paths so the move stays behavior and lint neutral

First colocation PR; establishes the _components private-folder convention for the remaining pages
