+++ ui/litellm-dashboard/eslint-suppressions.json
@@ -505,7 +505,7 @@
       "count": 1
     }
   },
-  "src/components/SearchTools/CreateSearchTools.tsx": {
+  "src/app/(dashboard)/search-tools/_components/CreateSearchTools.tsx": {
     "no-restricted-imports": {
       "count": 1
     },
@@ -513,17 +513,17 @@
       "count": 1
     }
   },
-  "src/components/SearchTools/SearchToolTester.tsx": {
+  "src/app/(dashboard)/search-tools/_components/SearchToolTester.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/components/SearchTools/SearchToolView.tsx": {
+  "src/app/(dashboard)/search-tools/_components/SearchToolView.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/components/SearchTools/SearchTools.tsx": {
+  "src/app/(dashboard)/search-tools/_components/SearchTools.tsx": {
     "no-restricted-imports": {
       "count": 1
     },
