+++ ui/litellm-dashboard/src/app/(dashboard)/search-tools/_components/SearchToolTester.tsx
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import { Button, Input, Typography, Spin } from "antd";
 import MessageManager from "@/components/molecules/message_manager";
 import { SearchOutlined, LoadingOutlined } from "@ant-design/icons";
-import { searchToolQueryCall } from "../networking";
-import NotificationsManager from "../molecules/notifications_manager";
+import { searchToolQueryCall } from "@/components/networking";
+import NotificationsManager from "@/components/molecules/notifications_manager";
 import { Card, Title as TremorTitle } from "@tremor/react";
 
 const { Text } = Typography;
