+++ ui/litellm-dashboard/src/app/(dashboard)/search-tools/_components/index.tsx
@@ -1,0 +1,1 @@
+export { default as SearchTools } from "./SearchTools";
