--- ui/litellm-dashboard/src/components/SearchTools/index.tsx
@@ -1,5 +1,0 @@
-export { default as SearchTools } from "./SearchTools";
-export { SearchToolView } from "./SearchToolView";
-export { default as SearchConnectionTest } from "./SearchConnectionTest";
-export { SearchToolTester } from "./SearchToolTester";
-export * from "./types";
